Upgrade Gatsby to v 3.5.1
fix package.json files

lock updates

remove unused lock file

more updates

more updates
