+++ @ndlib/gatsby-source-appsync-marble/gatsby-node.js
@@ -49,3 +49,82 @@ exports.sourceNodes = async (
     }
   }
 }
+
+// predefine stuff we expect from configuration.js
+exports.createSchemaCustomization = ({ actions }) => {
+  const { createTypes } = actions
+
+  const typeDefs = `
+  ### Marble Item Stuff
+  type metadataData @dontInfer {
+    label: String
+    value: [String]
+    urlField: String
+    type: String
+  }
+  type marbleIiifFile @dontInfer {
+    default: String
+    service: String
+    thumbnail: String
+  }
+  type MarbleFile implements Node {
+    id: String!
+    marbleId: String!
+    collection: MarbleItem
+    parentId: String
+    name: String
+    title: String
+    sequence: Int
+    file: String
+    extension: String
+    fileType: String
+    iiif: marbleIiifFile
+    local: File @link(by: "name", from: "name")
+    marbleParent: MarbleItem @link(by: "id", from: "parentId")
+  }
+  type MarbleItem implements Node {
+    id: String!
+    marbleId: String!
+    slug: String!
+    display: String
+    title: String!
+    description: String
+    sequence: Int
+    iiifUri: String
+    partiallyDigitized: Boolean
+    metadata: [metadataData]
+    copyrightRestricted: Boolean
+    childrenMarbleItem: [MarbleItem]
+    childrenMarbleFile: [MarbleFile] @link(by: "parentId", from: "id")
+    citation: String
+    parentId: String
+    marbleParent: MarbleItem @link(by: "marbleId", from: "parentId")
+    searchData: searchData
+  }
+
+  type searchData {
+    id: String
+    name: String
+    creator: [String]
+    collection: [String]
+    parent: String
+    identifier: [String]
+    geographicLocation: [String]
+    repository: String
+    themeTag: [String]
+    expandedThemeTag: [String]
+    centuryTag: [String]
+    date: String
+    lowestSearchRange: Int
+    highestSearchRange: Int
+    workType: [String]
+    thumbnail: String
+    language: [String]
+    type: String
+    url: String
+    formatTag: [String]
+    allMetadata: [String]
+  }
+  `
+  createTypes(typeDefs)
+}
