+++ @ndlib/gatsby-source-appsync-marble/package.json
@@ -15,5 +15,8 @@
   "dependencies": {
     "batch-promises": "^0.0.3",
     "isomorphic-fetch": "^3.0.0"
+  },
+  "peerDependencies": {
+    "gatsby": "^3.6.1"
   }
 }
