+++ @ndlib/gatsby-theme-marble/package.json
@@ -6,103 +6,104 @@
   "author": "Web and Software Engineering at Hesburgh Libraries",
   "main": "index.js",
   "dependencies": {
-    "@babel/core": "^7.10.2",
+    "@babel/core": "^7.14.3",
     "@babel/helper-skip-transparent-expression-wrappers": "^7.12.1",
-    "@blueprintjs/core": "^3.31.0",
-    "@blueprintjs/icons": "^3.20.1",
-    "@emotion/core": "^10.0.15",
-    "@emotion/styled": "^10.0.15",
+    "@blueprintjs/core": "^3.44.3",
+    "@blueprintjs/icons": "^3.26.1",
+    "@emotion/core": "^11.0.0",
+    "@emotion/styled": "^11.3.0",
     "@mdx-js/react": "^1.3.1",
-    "@okta/okta-auth-js": "4.8.0",
-    "@theme-ui/preset-bootstrap": "^0.3.5",
+    "@theme-ui/preset-bootstrap": "^0.9.1",
     "@types/googlemaps": "^3.39.6",
     "@types/markerclustererplus": "^2.1.29",
+    "autoprefixer": "^10.2.5",
     "babel-eslint": "^10.0.2",
-    "babel-jest": "^26.0.1",
-    "babel-preset-gatsby": "^0.4.9",
+    "babel-jest": "^27.0.1",
+    "babel-preset-gatsby": "^1.5.0",
     "crypto": "^1.0.1",
     "fetch": "^1.1.0",
-    "gatsby": "^2.32.4",
-    "gatsby-cli": "^2.7.30",
-    "gatsby-graphiql-explorer": "^0.2.3",
-    "gatsby-image": "^2.2.8",
-    "gatsby-link": "^2.2.4",
-    "gatsby-node-helpers": "^0.3.0",
-    "gatsby-page-utils": "^0.0.5",
+    "gatsby": "^3.6.1",
+    "gatsby-cli": "^3.5.0",
+    "gatsby-graphiql-explorer": "^1.5.0",
+    "gatsby-image": "^3.5.0",
+    "gatsby-link": "^3.5.0",
+    "gatsby-node-helpers": "^1.2.1",
+    "gatsby-page-utils": "^1.5.0",
     "gatsby-plugin-accessibilityjs": "^1.0.3",
-    "gatsby-plugin-image": "^1.0.0",
+    "gatsby-plugin-image": "^1.5.0",
     "gatsby-plugin-next-seo": "^1.7.0",
-    "gatsby-plugin-node-fields": "^2.0.1",
-    "gatsby-plugin-offline": "^2.2.5",
-    "gatsby-plugin-page-creator": "^2.1.16",
-    "gatsby-plugin-react-helmet": "^3.1.3",
-    "gatsby-plugin-robots-txt": "^1.5.0",
-    "gatsby-plugin-sharp": "^2.14.1",
-    "gatsby-plugin-sitemap": "^2.2.6",
-    "gatsby-plugin-theme-ui": "^0.8.0",
-    "gatsby-plugin-typography": "^2.3.2",
-    "gatsby-react-router-scroll": "^2.1.3",
-    "gatsby-source-filesystem": "^2.1.9",
-    "gatsby-transformer-json": "^2.4.5",
-    "gatsby-transformer-sharp": "^3.0.0",
+    "gatsby-plugin-node-fields": "^3.1.0",
+    "gatsby-plugin-offline": "^4.5.1",
+    "gatsby-plugin-page-creator": "^3.5.0",
+    "gatsby-plugin-react-helmet": "^4.5.0",
+    "gatsby-plugin-robots-txt": "^1.6.2",
+    "gatsby-plugin-sharp": "^3.5.0",
+    "gatsby-plugin-sitemap": "^4.1.0",
+    "gatsby-plugin-theme-ui": "^0.9.1",
+    "gatsby-plugin-typography": "^3.5.0",
+    "gatsby-react-router-scroll": "^4.5.0",
+    "gatsby-source-filesystem": "^3.5.0",
+    "gatsby-transformer-json": "^3.5.0",
+    "gatsby-transformer-sharp": "^3.5.0",
+    "graphql": "^15.5.0",
     "gravatar": "^1.8.0",
-    "i18next": "^19.4.5",
-    "js-file-downloader": "^1.1.7",
+    "i18next": "^20.2.4",
+    "js-file-downloader": "^1.1.19",
     "lodash.merge": "^4.6.2",
     "md5": "^2.2.1",
-    "mirador-image-tools": "^0.8.0",
     "moment": "^2.24.0",
+    "process": "^0.11.10",
     "prop-types": "^15.7.2",
-    "query-string": "^6.4.2",
-    "react": "16.11",
-    "react-beautiful-dnd": "^11.0.5",
+    "query-string": "^7.0.0",
+    "react": "17.0.2",
+    "react-beautiful-dnd": "^13.1.0",
     "react-dev-utils": "^11.0.4",
-    "react-dom": "16.11",
-    "react-google-maps": "^9.4.5",
-    "react-helmet": "^5.2.0",
+    "react-dom": "17.0.2",
+    "react-helmet": "^6.1.0",
     "react-helmet-async": "^1.0.7",
     "react-hot-loader": "^4.12.11",
     "react-i18next": "^11.5.0",
-    "react-markdown": "^4.0.8",
+    "react-markdown": "^6.0.2",
     "react-modal": "^3.11.2",
-    "react-reconciler": "^0.21.0",
+    "react-reconciler": "^0.26.2",
     "react-redux": "^7.0.2",
-    "react-select": "^3.0.4",
+    "react-select": "^4.3.1",
     "react-sortable-dnd-list": "^1.0.1",
     "react-typography": "^0.16.19",
     "redux": "^4.0.1",
     "redux-thunk": "^2.3.0",
-    "searchkit": "2.4.1-alpha.4",
-    "theme-ui": "^0.6.0-alpha.7",
-    "typography": "^0.16.19",
+    "searchkit": "2.4.4",
+    "theme-ui": "^0.9.1",
+    "typography": "^0.16.21",
     "typy": "^3.1.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.0.0",
-    "@typescript-eslint/eslint-plugin": "^3.2.0",
-    "@typescript-eslint/parser": "^3.2.0",
+    "@typescript-eslint/eslint-plugin": "^4.24.0",
+    "@typescript-eslint/parser": "^4.24.0",
     "babel-eslint": "^10.0.2",
     "babel-jest": "^26.0.1",
-    "babel-preset-gatsby": "^0.4.9",
+    "babel-preset-es2015": "^6.24.1",
     "enzyme": "^3.9.0",
     "enzyme-adapter-react-16": "^1.12.1",
     "eslint": "^7.2.0",
-    "eslint-config-react-app": "^5.0.0",
-    "eslint-config-standard": "^14.1.1",
-    "eslint-config-standard-react": "^9.2.0",
+    "eslint-config-react-app": "^6.0.0",
+    "eslint-config-standard": "^16.0.2",
+    "eslint-config-standard-react": "^11.0.1",
     "eslint-plugin-flowtype": "^5.1.3",
     "eslint-plugin-import": "^2.21.1",
     "eslint-plugin-jsx-a11y": "^6.2.3",
     "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-promise": "^4.2.1",
+    "eslint-plugin-promise": "^5.1.0",
     "eslint-plugin-react": "^7.20.0",
     "eslint-plugin-react-hooks": "^4.0.4",
-    "eslint-plugin-standard": "^4.0.0",
+    "eslint-plugin-standard": "^5.0.0",
     "gatsby-plugin-root-import": "^2.0.5",
     "identity-obj-proxy": "^3.0.0",
-    "jest": "^26.0.1",
+    "jest": "^27.0.1",
     "jest-environment-enzyme": "^7.0.2",
     "jest-enzyme": "^7.0.2",
+    "path": "^0.12.7",
+    "path-browserify": "^1.0.1",
     "prettier": "^2.0.5"
   },
   "keywords": [
