+++ @ndlib/gatsby-theme-marble/src/components/Shared/ActionButtonGroup/ActionButton/index.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import style from './style.module.css'
+import * as style from './style.module.css'
 
 const ActionButton = ({ name, action, icon, activeIcon, isActive, altText }) => {
   return (
